Résolution bugs

+++ E-Health/src/enseirb/t3/e_health/process/DataProcess.java
@@ -23,21 +23,21 @@ public class DataProcess {
 		case "A":
 			break;
 		case "B":
-			if (Integer.parseInt(data.getValue()) > 100)
+			if (Double.parseDouble(data.getValue()) > 100)
 				alertes.add("Tachycardie");
-			else if (Integer.parseInt(data.getValue()) < 60)
+			else if (Double.parseDouble(data.getValue()) < 60)
 				alertes.add("Bradycardie");
 			break;
 		case "O":
-			if (Integer.parseInt(data.getValue()) < 90)
+			if (Double.parseDouble(data.getValue()) < 90)
 				alertes.add("Hypoxémie");
 			break;
 		case "P":
 			break;
 		case "T":
-			if (Integer.parseInt(data.getValue()) > 38)
+			if (Double.parseDouble(data.getValue()) > 38)
 				alertes.add("Hyperthermie");
-			else if (Integer.parseInt(data.getValue()) < 35)
+			else if (Double.parseDouble(data.getValue()) < 35)
 				alertes.add("Hypothermie");
 			break;
 		default :
